migrate redis client
