+++ packages/backend/src/persistance/messages/createMessage.ts
@@ -1,6 +1,5 @@
 import * as Lib from 'dm3-lib/dist.backend';
-import { RedisPrefix } from '../../redis';
-import { Redis } from '../getDatabase';
+import { Redis, RedisPrefix } from '../getDatabase';
 
 export function createMessage(redis: Redis) {
     return async (
