+++ packages/backend/src/persistance/pending/deletePending.ts
@@ -1,8 +1,7 @@
-import { RedisPrefix } from '../../redis';
-import { Redis } from '../getDatabase';
+import { Redis, RedisPrefix } from '../getDatabase';
 import * as Lib from 'dm3-lib/dist.backend';
 
-export async function deletePending(redis: Redis) {
+export function deletePending(redis: Redis) {
     return async (accountAddress: string): Promise<void> => {
         await redis.del(
             RedisPrefix.Pending + Lib.external.formatAddress(accountAddress),
