+++ packages/backend/src/persistance/storage/setUserStorage.ts
@@ -1,6 +1,5 @@
 import * as Lib from 'dm3-lib/dist.backend';
-import { RedisPrefix } from '../../redis';
-import { Redis } from '../getDatabase';
+import { Redis, RedisPrefix } from '../getDatabase';
 
 export function setUserStorage(redis: Redis) {
     return async (accountAddress: string, data: string): Promise<void> => {
