--- packages/backend/src/redis.ts
@@ -1,38 +1,0 @@
-import 'dotenv/config';
-import { createClient } from 'redis';
-import * as Lib from 'dm3-lib/dist.backend';
-import { Express } from 'express';
-import { stringify } from 'safe-stable-stringify';
-
-const endpointUrl = process.env.REDIS_URL || 'redis://127.0.0.1:6379';
-
-export enum RedisPrefix {
-    Conversation = 'conversation:',
-    Sync = 'sync:',
-    Session = 'session:',
-    UserStorage = 'user.storage:',
-    Pending = 'pending:',
-}
-
-export async function createRedisClient(app: Express) {
-    const socketConf = {
-        socket: {
-            tls: true,
-            rejectUnauthorized: false,
-        },
-    };
-    /*     const client = createClient({
-        url: endpointUrl,
-        ...(process.env.NODE_ENV == 'development' ? {} : socketConf),
-    }); */
-
-    const client = createClient();
-    client.on('error', (error) => {
-        app.locals.logger.error({
-            method: 'REDIS CLIENT',
-            error,
-        });
-    });
-    await client.connect();
-    return client;
-}
